feat: 优化模版
