+++ packages/gi-assets-advance/src/components/CypherQuery/CypherCodeMirror/codemirror-cypher.css
@@ -1,3 +1,8 @@
-@import "./css/editor.css";
-@import "./css/syntax.css";
-@import "./css/completion.css";
+@import './css/editor.css';
+@import './css/syntax.css';
+@import './css/completion.css';
+
+.CodeMirror {
+  min-height: 180px !important;
+  height: 100% !important;
+}
