+++ packages/gi-assets-advance/src/components/CypherQuery/index.less
@@ -1,6 +1,7 @@
 .cypher-query-container {
   position: relative;
   padding: 16px;
+  height: 100%;
   .publishButton {
     position: absolute;
     right: 118px;
